Fix join
This join code was doing nothing due to the typo.
As a result the task was trying to switch the license
of adopted sentences. However the Sentence model has
validation rules that did not allow such license switch,
so the license switching process was working and the test
was passing. In addition, the error was not reported
because error reporting was not migrated to CakePHP 3
and that code is not covered by tests.