Revert use of CounterCache
Refs #1774.

I have a feeling that it’s not a good idea to add a new lang field
to the audios table, because it is against the principle of single
source of truth. It’s simpler to let the framework handle the count,
but not to the expense of adding unnecessary fields to the schema.