Specify binding key for groups table relation
This commit allows to write tests that involve saving data
in the users table without having to unnecessarily include
the groups table in the fixtures.

We’re having a rule that checks if a given value for
users.group_id matches an existing group in the groups table.
This is all good and well, but when writing a test that saves
data in the users table, CakePHP tries to fetch from the groups
table even though the field users.group_id is untouched. This
is because CakePHP checks what is the primary key of the groups
table before even checking if the users.group_id is being
updated. By specifying the binding key, CakePHP won’t be
touching the groups table unless some users.group_id is modified.

Below is a typical error that this commit prevents:

PDOException: SQLSTATE[42S02]: Base table or view not found: 1146 Table 'tatoeba_test.groups' doesn't exist

/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Database/Statement/MysqlStatement.php:39
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Database/Connection.php:350
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Core/Retry/CommandRetry.php:68
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Database/Connection.php:353
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Database/Connection.php:302
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Core/Retry/CommandRetry.php:68
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Database/Connection.php:306
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Database/Schema/Collection.php:129
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Database/Schema/Collection.php:97
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Database/Schema/CachedCollection.php:63
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/ORM/Table.php:555
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/ORM/Table.php:680
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/ORM/Association.php:585
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/ORM/Rule/ExistsIn.php:98
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Datasource/RuleInvoker.php:115
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Datasource/RulesChecker.php:298
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Datasource/RulesChecker.php:268
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Datasource/RulesChecker.php:235
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Datasource/RulesAwareTrait.php:66
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/ORM/Table.php:1982
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/ORM/Table.php:1921
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/ORM/Table.php:1634
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/Database/Connection.php:738
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/ORM/Table.php:1635
/home/vagrant/Tatoeba/vendor/cakephp/cakephp/src/ORM/Table.php:1922
/home/vagrant/Tatoeba/src/Model/Licensing.php:14
/home/vagrant/Tatoeba/src/Model/Licensing.php:30
/home/vagrant/Tatoeba/src/Model/Licensing.php:48
/home/vagrant/Tatoeba/tests/TestCase/Model/LicensingTest.php:67