Update cron files for Manticore
With Sphinx, we had our own .service file that used to run
searchd with --console. This options prevents Sphinx from
creating a PID file, so we needed to set the PID file manually.
Now we use manticore’s .deb and it comes with a .service file
that does not run searchd with --console.

Refs #1518.