Refactor Sphinx options calculation method
Not only the code is clearer, but it also solves a weird bug
when $query is empty and the sort order is "Last modified first"
and "Reverse order" is not checked. In that specific case,
it looks like Sphinx orders on the "modified" attribute while
ignoring some of the least significant bits of the attribute.