Fix tests for resetting script fields
These tests were passing on Jenkins because
AutoTranscriptions.enabled is set to false, so it ignored
the additional checks in TranscriptableBehavior, like
isValidScriptForLanguage(), that would have prevented the
test from passing because 'TEST' is not a valid script for
language 'cmn'.

However we want AutoTranscriptions.enabled to be enabled
because it doesn’t make sense to run this shell without it.