Display server-side validation error within form
for the 'username' field.

Note that I couldn’t find a better way than setTimeout()
to make the validation error appear on page load. Even
after invalidating the field, the error is still hidden in
the background, waiting to show up with the help of the
animation thing. The only way to trigger that animation is
to change the "$touched" property of the field (or to
manually click on it).

Refs #1270.