Fix route problems during tests
These tests are playing with UI.languages, which is used in
config/routes.php for routes definition. However when
caching is enabled, this temporary value for UI.languages
that we use in AppControllerTest can end up being cached,
and can then make other tests failing, like
AppControllerTransitionTest.php because it accesses /fra.