Fix test checking redirection to last page
Fixed test WallControllerTest::testPaginateRedirectsPageOutOfBoundsToLastPage()
that broke because of a8d07bbad919221a527a8b0d8d8b3cb86e5def98.
AppController::paginate() was returning array(),
whereas we assumes it returns a ResultSet.

It’s not a good idea to redirect from inside paginate(),
because the proper way of redirecting is to `return` a new
Response object from the controller action. This makes
the code match better what happens in reality.