Serve the source code over NFS
Note that Vagrant’s synched folders are all meant to make
a directory on the host available to the guest, whereas we
want the opposite. The reason we are doing this is that we
want to provide VMs that contain the source code, so that
we are sure they are working, no matter how old the code is.