Add translations in list exports
I had to rework the query to use ->matching(), so that when
a sentence has several translations, we get one result per
translation, as opposed to one entity with all linked
translations attached. Using ->matching() also simplifies
the ->select() since we only need to provide the id for the
batch operation to work.