Pagination only works reliable with distinct values
This pull request addresses issue #1594.

Pagination depends on a stable result set because it queries the
database several times in order to split the result in pages. But
`view_all()` in the `Tags` controller uses `nbrOfSentences` as the
ordering which isn't stable because several tags could have the same
number of sentences linked to them and MySQL doesn't guarantee the
order when there is a tie.

Using `id` instead ensures that there will always be a tiebreaker
because its values are distinct.