Work around potential race condition
See bug #1944. I suspect this bug is caused by a race condition
between index merge and refresh. This commit is an attempt to
prevent that race condition from happening, in order to check
if it is actually causing bug #1944.

By using the --sighup-each flag, when an index merge starts,
the first indexes of the list will have been rotated already.