Simplify ago() function in Date Helper
Instead of setting the format for dates at several places we set it
once in bootstrap.php. This also allows us to use nice() from the
Time Helper for formatting the date.

Additionally, use parseDateTime() instead of createFromFormat()
for parsing the given date string. By default, parseDateTime() uses
the "toString" format which we also set in bootstrap.php.