Improve handling of 'datetime' database columns
In most parts we still used strings for the content of 'datetime'
columns. But in combination with the Timestamp behavior and locales
which don't use arabic numerals (in our current UI Arabic, Bangla and
Marathi) this is broken because CakePHP internally represents
Timestamp dates as FrozenTime objects and when these get converted
to strings, the locale numerals are used. MySQL can't use these
strings and as a consequence the dates are stored as
'0000-00-00 00:00:00' in the database.

Issue 1948 demonstrated this problem.

I've changed all problematic tables to use FrozenTime objects for
representing dates. There are still tables which use strings for dates
(mainly PrivateMessages and the tables handling contributions) but
since we currently don't use the Timestamp behavior there they still
work and I left them as is (it's probably a good idea to change them in
the near future).

I've also added tests and as a bonus fixed some function documentation
comments. :-)