Fix HashTrait
The initial version of HashTrait had the major flaw that hash values
from the database where always recalculated irrespective of whether they
were correct or wrong.

This version gets the actual hash from the database and only updates the
hash when at least one of the fields the hash is based on is modified.
This allows us to easily check the correctness of the stored hashes.