Add Cake Command for fixing tables with wrong hashes
The command basically checks every row in a given table and compares the
stored hash with a new calculated hash based on the fields the hash is
based on. It then updates the hash column if necessary.

While testing the command with the test Sentences table I've discovered
3 rows with wrong hashes (in addition to the two rows I've modified
  intentionally). So it seems to work :-)

`cake fix_hashes --help` provides more usage information.