Optimize paginateLatest
For calculating the lowest id we only need the associations used in the
'select' and 'order' part of the query.

Since CakePHP eagerly loads all the associations for a table, using the
original 'contain' part of the query results in unnecessary database
queries.

For an unregistered user, this new version reduces the number of queries
sent to the database from 13 to 9 for the 'Browse by language' page. For
a registered user the number of queries is reduced from 26 to 16 with
the new design and from 56 to 46 with the old design.