Remove AUTO_INCREMENT table option in some table creation scripts
Some table creation scripts have a AUTO_INCREMENT option which is
unnecessary and could hurt us whenever we update MariaDB to 10.2.4 or
later.

With our current version (10.1.38), the counter for the autoincrement is
stored in memory. Thus after every restart of the database server (as we
do when creating the tables using imouto) the counter is re-initialized
to the highest value found in the table.

But 10.2.4 changed this behavior and the counter is persistent. As a
consequence any table creations after a MariaDB update would start with
the value defined in the AUTO_INCREMENT option.

Since we want our new tables to always start at 1, I simply removed the
AUTO_INCREMENT option.