Significantly improve homepage load time
$this->Sentences->find()->count() is actually a very costly request
because it takes seconds to goes through the whole sentences table
just to count the total number of rows. We do cache the stats block
to avoid recalculating that, but we actually execute the SQL request
every single time, only to update the cached stats block every
15 minutes. This commit makes it so that the SQL request is only
executed when the stats block is updated.