FixHashesCommand: Added option for reading ids from stdin
In addition this new version also checks whether any of the given
columns changed after creating a new entity from the old values. If this
is the case, all columns are updated and not only the hash column.

As a result it is now possible to fix specific entities in cases where
any of the source columns is updated with new contstraints (e.g.
normalizing Unicode text).