Avoid dying in the middle of the test suite
If I start the whole test suite and go make some tea while
it’s running, I don’t want it to stop in the middle just
because a particular test failed. Besides, failure to
create a directory will trigger a PHP warning, which is
enough to figure out what went wrong.