Replace embedded newlines with spaces in exports
Currently the text of sentence comments and wall posts is exported as it is
stored, i.e. with embedded newlines. This makes it harder to process the data
with other tools (e.g. importing it into a local database).

Replacing the embedded newlines with spaces ensures that every row in
the table is represented by a single line in the downloadable file.