Fix "Owned by a self-identified native" criterion
Closes #2166.

Work around Manticore filter values limitation.
When the number of natives is greater than 4096, Manticore
throws an error. To avoid this, we filter by excluding
non-natives instead. This is possible because filters are
combined with a boolean AND operation, so we can create
multiple filters with 4096 values each.

At the moment, on Tatoeba, there are 4128 English natives
and 5129 non-natives.