Add a new better script to update transltions
Replace the old script with a new one based on the
Transifex client. The new script does the same thing
with less code. In addition, the new script also:

* downloads in parallel
* retreives the Terms of Service translations
* retreives languages that are not listed in lang_map

The -a option is meant to be used on dev.tatoeba.org only.

lang_map is now only useful for languages that use a
different directory name than the Transifex language code.