Make trigger on contributions_stats fill more cases.
Add the sentence+delete, link+insert, and link+delete cases.
It ignores the cases when action="update" and type="license".
When the deleted sentence is one that was added today, it decreases the 
"insert" counter. Otherwise, it increases the "delete" counter.
The same mechanism is not implemented for linking/unlinking.