Fix 'Unknown provider' error on homepage for guests
initLists() will inject the listDataService, but the javascript will not be loaded when the user is not authenticated.

We have to removed the initList() from init() because init() is executed for both authenticated and non-authenticated users. It has to be added in the listener for 'randomSentenceRequested' instead, so that the list choices are properly set when authenticated users are loading a new random sentence.