Add per-language selection to some sections on Downloads page
The user can choose between downloading sentences in all languages or
only the sentences in the selected language.

I've also noticed that for a logged-in user the initial version which
used Languages->unknownLanguagesArray() doesn't work because the list is
splitted in preferred languages and other languages.

I've also changed the behaviour when there are no per-language files
available.