Fix new terms of service acceptance
Users who modified their settings (but who were never prompted
to accept the new Tos because of the bug resolved in
1b243fc81ceaa42a525d85bbdf6e9b0a2ee4bb51) have new_terms_of_use
set to false, so we now need to fix that setting for them.

Users who never changed their settings don’t need any update
because the default value is read from the source code, which
is now correct.

Note that I used the value '1' for new_terms_of_use because
that is what users who accepted to switch to the new ToS have
in their settings. I thought we may use other values (instead
of true/false) for future ToS revisions.