Avoid xss attacks when initializing AngularJS models
Whenever we initialize a model with user input using 'ng-init' we need
to escape the opening curly brackets in order to avoid template
injection attacks.

In addition we should create the input forms manually instead of using
CakePHP's FormHelper because it will prefill the 'value' input with the
unsafe user input.