Refactor LicenseHelper
We used the LicenseHelper in two different configurations: one for
valid sentence licenses and one for valid audio licenses.

But this was a problem if we had to use both configurations on one page
(e.g. the sentence page). There was a hack which solved the problem by
unloading the first configuration before the second configuration was
used but this isn't a good solution.

So I've put all the license data into a Lib class which is now easily
available everywhere (similar to the languages and countries lists).
In addition I've split the LicenseHelper into three parts: LicenseHelper,
which contains all the general parts, and the two children
SentenceLicenseHelper and AudioLicenseHelper which define the valid
licenses for each context.