Change unlink icon
The scissors look too big and look like they are cutting the flag, which is not properly conveying the action of unlinking.

Using `link_off` icon instead: https://material.io/resources/icons/?icon=link_off&style=baseline
Since this icon is not part of the Material icon font, we need to include it manually. We may want to do this for every <md-icon> to solve #1985.