Hide "@translators" markers in POT file
And separate multiple @translators phrases with slash.

There are no line returns in developers notes embedded
in POT files, everything is just concatenated, therefore
we need to do some formatting so that the resulting note
on Transifex is more or less readable by translators.

In other words this commit changes the developers notes
in the POT files from this:

  #. @translators: title of the help text on the Adopt sentence page
  #. @translators: title of the help text on the Add vocabulary page
  #. @translators: title of the sidebar text on the Wall page

into this:

  #. title of the help text on the Adopt sentence page
  #. / title of the help text on the Add vocabulary page
  #. / title of the sidebar text on the Wall page

Which is seen on Transifex as:

title of the help text on the Adopt sentence page / title of the help text on the Add vocabulary page / title of the sidebar text on the Wall page