Avoid inserting \r in developers notes
When the source file is encoded using \n\r line returns,
and contains a multiline developers note, such as in
src/Template/Activities/translate_sentences.ctp, the \r
character gets into the note in the final POT file.

We want to avoid that so that line returns are consistent
in the POT file.