Avoid unnecessary calls to the search engine
CakePHP's paginator sends two queries to the database: the first for getting
the items of a page and the second for the total count.

But when we use the search engine we get the total count already with
the first call. The 'withSphinx' finder from the Sentences model already
uses the cached result for the total count but we still queried the
search engine twice in the Spinx behavior.