Fix licences bug in add sentence form (new design)
If a user who cannot switch the license used the form for adding a new
sentence in the new design, the sentence was submitted with the
incorrect CC0 license.
The license for a new sentence was initialized with the first entry
from the license array which happened to be the wrong license. After
removing this line, the sentence will be submitted without a license and
thus it gets assigned the user's default license in the model.