Remove unnecessary assertion
Upgrading from PHP 7.0 to 7.3 produces the following warning:

Warning Error: count(): Parameter must be an array or an object that implements Countable in [/home/vagrant/Tatoeba/tests/TestCase/Model/Table/SentencesTableTest.php, line 373]

That assertion is unnecessary because we can be sure that
$transcrBefore and $transcrAfter will always contain exactly
one result. If not, the ->find()->first() calls that assign
them would have raised an exception.