Fix caching the Manticore query
Using `$this->_cached_result` only works if there is just one query per
request. But on the vocabulary page we send several queries in one
request and thus we always got back the same result.
We should instead check whether the new query is exactly the same as
the cached one.