Allow to set random seed on search form
I experimentally tested the behavior of the seed parameter and
concluded that it’s a 64-bit integer. However, we don’t really
need that much randomness for sorting results. Also, such a huge
value would take a lot of space in the search URL, which is
already quite long enough with all its other parameters.

So I figured out 24 bits nicely fits in 4 characters of base64,
and it should be big enough for random sorting.