Make rand_seed redirection URL nicer
When performing a search in random order, if the rand_seed
parameter is not provided, we redirect to the same URL including
a new rand_seed parameter. That URL, generated by the
Router::url() method, should be as close as possible to the
original requested URL. That’s why we want to keep the same
array order.