Redesign top menu to for small screens
When the UI language label is too long, it triggers horizontal scrolling and some important elements of the UI (such as the button to expand the advanced search) can be hidden. To ensure the elements in the top menu are not too wide, we don't display the language name on small screens, just the icon.

Additionally, for medium screens and below, we move the "Register" and "Login" items into the hamburger menu to avoid taking unnecessary vertical space.