Really fix audio in the old design
1df8c35c61a42757abfdc80cee92c1d7e1cf8f7f only works when the audio was
already known on page load. But on pages where a sentence with audio was
dynamically loaded (e.g. a duplicate sentence on the 'Add sentences'
page) the audio wasn't available.

This commit simply adds an audio element whenever we display the audio
button and so doesn't need to modify the audio button on the client-side in
order to play the audio.