Revert removing jQuery from layout.js
dcdab7f was to optimistic in the assumption that jQuery is only needed
for displaying the full sentence block (with all the menu buttons). But
we have still pages where we display a sentence in a minimal version of
the old design (just the sentence + the copy and audio button). These
pages (e.g. all the audio pages) still require jQuery (and `watch.js`)
so we need to load it even if we don't need `sentences-block-for-members.js`

So this commit reverts parts of dcdab7f and the unsuccesful attempts to
fix regressions caused by it (c8cc8aa and 1df8c35).