Prevent debug from messing up with Json responses
When a Json response is to be sent and debug is enabled, then,
if some PHP notice/warning is produced, an HTML stacktrace is
inserted into the response, effectively corrupting the Json
code. This can break the front-end even if the warning is
harmless. This commit prevents the insertion of that HTML code
inside Json, but the notice/warning is still written to
debug.log.

An example of this scenario is the following notice produced
from /sentences/add_an_other_sentence when adding a sentence
into a language that is not configured in Sphinx:

  Unable to update Sphinx attribute(s) (user_id, ucorrectness):
  searchd error: unknown index 'lfn_delta_index'

Besides, as we are rewriting more and more front-end code into
AngularJs, this can only higher the chances of getting such
scenarios.