Work around Firefox rendering bug on SVG flags
The drop-shadow() filter that we use on all flags makes them
blurry when the flag is displayed as a <use> SVG image on
Firefox. To work around this problem, we replace it with
box-shadow, which is mostly okay but doesn’t display nicely
on non-rectangular flags such as the Marathi one.