Optimize internal query of 'latest' custom finder
When ->find('latest') is used along with ->find('nativeMarker'),
it runs its internal query with an unnecessary join.
We can assume that all LEFT joins are unnecessary because they
do not affect the number of results (as long as they do not
bring new fields used in the main query).