Fix wrong license for translations
The license for a translation was never explicitly set and so it was set
in SentencesTable::beforeSave to the user's default license which in the
past was always 'CC BY'. But c5191c87e124079dfb3b27fe2989029ba27c3f72
enabled users to change their default license in the settings.
So when a user changed their default license to 'CC0', the license of a
translation was always incorrectly set to that license.