Allow tags.category_id to be NULL
When adding the column, we don’t know yet what the value
category will be, so we need to set it to NULL. If this is
not allowed, mysql will set the value to zero instead and
we don’t want that.

If you already executed this migration, just run

    cake migrations rollback && cake migrations migrate

to fix the column. You will loose existing assignment of
tags to categories.