Disallow /categories_tree/* on production for now
The way we are going to create the tree is still unclear.
Giving read-write access to the tree to any advanced user
and above, like for tags, is perfect for testing out, but
it’s definitely not the way we are going to create the real
tree. We need to think this through with the community
first.

Refs #2435.