Fix pagination on Search page
In f815c25 we use a query instead of the model for pagination. The query
uses `find()` for setting custom finders instead of the `finder` option of
`$paginate`. But these finders now run before the paginator who adds an
offset clause to the query.
We should remove the offset (and also the unnecessary limit) clause in
SphinxBehavior's `beforeFind` method.