Fix display of long language names
In src/Template/Element/sentences/index/all.ctp, when
setting the UI in Finnish, some language names are very
long and overflow over the flag of the adjacent column,
despite of our global "word-wrap: break-word" CSS rule,
because of our CSS "display: table". It looks like the
proper way to solve this is to set a max-width value.
However, I am unable to do this because the column width
is variable; we only know it’s at least 150px. To work
around this problem, I set "word-wrap: anywhere" there to
get the expected behavior of "wrap: break-word" in a
fixed-width block. I also set "hyphens: auto" so that, if
the browser supports it, the word breaking looks nicer.