Change locale for CLI utilities
The default locale is set to 'en_US' but the English web UI uses 'en'.
Thus CakePHP shells/commands created extra cache files for translations.

Configuring the language in the shell/command doesn't work, because the
code for setting the locale from the configuration language is in
Controller/AppController.php which isn't used by CLI tools.