Fix UI language preselection
Stop considering the "language in URL" as a legit base in
UI language preselection. It causes problems because users
get their UI language initially set because of something
out of their control: the language the link creator had
used.

For example, if I open for the first time
https://tatoeba.org/eng from search engine results, I would
end up on the English homepage regardless of my browser
language settings.

We already have a cookie to keep the language consistent
over a session, so there is no need to care about the
language in URL.

Refs #2544.