Make sidenav scrollable again on mobile
Fixes #2567.

`height: auto` on the body makes the height of the sidenav extend beyond the height of the browser window, and therefore removing the scroll. Since the need to scroll inside the sidenav only applies to medium and small screens, we can set a media query so that `height: auto` is not set for width below 960px.