Refactor SentencesTable::editSentence
When the client sent the data to the server, it combined the sentence id
and language into a single string. The probable reason for this is the
default configuration of the `jeditable` library which uses the id of the
surrounding `div` element for the `id` key of the POST data.

But it's also possible to format the POST data that is sent and we
should take advantage of it. This simplifies the code on the server (and
for the new design) and also makes `editSentence` easier to reuse.