Remove unused flag image files
`kur` (Kurdish) was removed in 936345f32 but I forgot to remove the flag
too
`add`, `chm` and `roa` were obviously mistakenly introduced in
977786c25:
`add` is the code for Dzodinka which we do not support and the flag
looks more like an "add" icon from a GUI.
`chm` is the code for Mari which we do not support (and which is
considered a macrolanguage).
`roa` is the code for the "Romance languages" collection.