Add new LanguageSelectorMiddleware
This moves the UI language selection part from the AppController to a
custom middleware. The advantage is that a possible redirect happens
directly after the router parsed the request and so we avoid starting
the full app.

The order for determining the UI language is now:
1) value in cookie
2) value from Accept-Language header
3) value from URL
4) English

This commit also changes the cookie for the interface language because
the old way doesn't work correctly anymore due to changes in how CakePHP
handles cookies in recent versions. (E.g. the Cookie::write() only
accepts just two arguments in CakePHP 3, so the way we set the
expiration time doesn't work anymore.)