Add email configuration for local email delivery
The Local transport will send all emails to the local mail
server. Logging of unsent emails in case of a problem can be
configured using the `log` key.
There is now also a default email profile which is used by CakePHP
whenever it creates a new Email object.
For the dev server the transport for this profile should usually be
set to `debug`. This replaces the old `Mailer.enabled = false` and
enables logging of all emails sent from the application.