Replace registration quiz with honeypot
This commit replaces the registration form quiz with a
honeypot (a hidden field spambots may be tempted to fill
in) because the quiz has multiple problems.

We sometimes receive emails from users complaining that
they cannot answer the quiz question correctly. The reason
it that they omit punctuation characters in their answer.
We already have an explanation text for that, but in the
real world people don’t read carefully, so that doesn’t
work.

In general, such captcha are not very UX-frienly because
they forces the user into performing a boring and seemingly
useless action.

While I don’t have any number to estimate the performance
of the quiz in preventing spambots from registering,
I believe it’s about the same as the honeypot (that is to
say pretty bad). But UX-wise the honeypot is definitely
better; that’s the real improvement this commit brings.