Suggest languages without clearing dropdown text
So far we have been clearing/restoring the text field of
language dropdowns on focus/blur because an empty field is
needed to get suggestions to appear in <md-autocomplete>.
However this is a problem UX-wise because it shows the
same results for two very differents actions: clearing
field and changing selection.

So I found a hack to make <md-autocomplete> display
suggestions on focus without emptying the field. My goal
is to get rid of the extremely long language list appearing
on focus, and instead only display priority languages. This
way, less useless information is displayed and the users
have an incentive to type in rather that going through that
very long list.