Bugfix in the suggestions hack
This commit fixes the following bug: when suggestions have
been triggered by focusing the field (with the hack) but
the clicked away without selecting a language, then the
language displays in red (validation problem). This is how
<md-autocomplete> works: as soon as the search text has
been edited, the language selection is lost, no matter if
we edit back the text to what it was before focus.