Fix focus bug around language dropdown "X" button
It looks like there is a kind of bug in <md-autocomplete>:
clicking within the 5px margin around the "X" clear button
triggers the onFocus() event without actually giving focus
to the input field. This looks harmless in a regular
md-autocomplete dropdown, however it breaks our suggestions
hack because suggestions are triggered without being
displayed. As a result the field turns red (validation
problem).