Append "show all" button in language suggestions
Provide a way to show *all* languages no matter what,
for users who cannot or don’t wanna use keyboard to
select a language that is not in suggestions list.

This is a pure hack that works because Material’s
autocomplete code apparently include the contents of
<md-item-template> without proper markup checking.
The code looks like that (edited for brievity):

  return '\
    <md-virtual-repeat-container>\
      <ul>\
        <li md-virtual-repeat="item in $mdAutocompleteCtrl.matches"\
            ng-click="$mdAutocompleteCtrl.select($index)">\
            ' + itemTemplate + '\
            </li>' + noItemsTemplate + '\
      </ul>\
    </md-virtual-repeat-container>'

So we can inject some extra <li>
by including a </li><li> sequence in itemTemplate.