Fix "clear button pressed" condition
If you click on a random button on the page, and then
click in the dropdown field, the suggestions are not shown
because the code thinks you pressed the X (clear) button.
We need not only to make sure that $event.relatedTarget
is a button, but also that it’s the dropdown X button.

See also cc0ffaef76895bfe7ce38e9cc888752413c582c5.