Fix language dropdown behavior in forms
Newer AngularJS Material forces a non-empty value for
the name="" attribute of the underlying <input> of
md-autocomplete. This enforces the submission of a field
having the value of the language name (as opposed to the
language code that we need).

To work around this, I assigned the same name="" attribute
to both the hidden input field containing the language
code and the language dropdown input containing the
language name. In addition, I put the hidden input at the
end so that its value overwrites the other input. This way,
the value we get in controllers is the same as before.

Refs #2561.