Make reset-button directive work with ng-model
The advanced search form now uses ng-model for all its
fields, and clearing the field by manipulating the DOM
directly does not trigger a model change so AngularJS
does not know about the cleared value. Which means
clearing the field and submitting the form right away
sent a non-cleared value.

Refs #2561