Remove exports page
I figured out it brings more confusion than any else to
show the fact that each user has a list of exports. The
typical use-case is to generate a export, download it and
forget it. The only useful reason we keep the file
available for some time is that it allows pausing and
restarting the download.

Refs #1609.