Improve merging of new translations
The old code assumed that we needed to update the .po file
using the .pot file as a reference. However, Transifex does
that already. What we need instead is to add new
translations to existing .po files, so that these can be
uploaded to Transifex. I removed the CLDR copyright header
because Transifex always removes it anyway.