Reorganize old design JS files
There was a regression that prevented japanese
transcriptions from working correctly in the old
design. In particular, on page
/sentences/of_user/xxx, guests could not see
furiganas as "blended" and members could not edit
them.

While fixing this, I grouped all the optional JS
files for "single line sentences" (as displayed
by SentencesHelper::displayGenericSentence())
into a new compressed file single-sentence.js.

It makes sense to group them because they are
likely to be required for many sentences. We have
many sentences with audio now, all requiering
sentences.play_audio.js. The copy button, using
sentence.copy.js, used to be opt-in, but it is
now displayed to everyone.