Change translations expand/collapse button label
Now that this button can hide *all* the
translations, it would not make sense to make it
say "show n more translations" in that case.

Also included the icon inside the translatable
string to allow translators to put it at a
different place (this should be useful for RTL
languages).