Fix PHP warning
If there is at least one sentence with unknown flag,
languagesHavingSentences() will include a NULL value to
account for that. We need to remove it from the result
because Manticore does not index sentences of unknown
language. The warning was:

Warning Error: array_flip(): Can only flip STRING and
INTEGER values! in [/home/vagrant/Tatoeba/src/Shell/
SphinxConfShell.php, line 750]