Revert adding column audios.audio_idx
We can simply use audios.id instead.

audios.id is not a human-friendly number, and I wanted to use a human-friendly
number to get nice file name for audio files, but it is not exposed to users
anyway. Power users who make use of scripts to download the audio already need
to rely on exported CSV sentences_with_audio.tar.bz2 to get the list of
sentences having audio, so we can just include audios.id in it and they’ll need
little modification to adapt their code.